requirements.txt
pyusb>=1.0.0,<2.0.0
pytest>=3.0.7,<4.0.0
